requirements.txt
bcrypt==3.2.0
bidict==0.21.2
cffi==1.14.5
click==8.0.1
Flask==2.0.0
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-SocketIO==5.0.3
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
peewee==3.14.4
psycopg2==2.8.6
pycparser==2.20
python-dotenv==0.17.1
python-engineio==4.2.0
python-socketio==5.3.0
six==1.16.0
Werkzeug==2.0.1
